requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
Automat==20.2.0
autopep8==1.5
awswrangler==1.4.0
backcall==0.2.0
beautifulsoup4==4.8.0
bleach==3.1.5
boto3==1.11.5
botocore==1.14.5
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.6.16
cffi==1.14.0
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
contextlib2==0.6.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.0.1
djangorestframework-jwt==1.11.0
docutils==0.14
EasyProcess==0.2.9
entrypoints==0.3
et-xmlfile==1.0.1
fsspec==0.6.2
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
html5lib==1.0.1
httplib2==0.18.1
hyperlink==19.0.0
idna==2.8
image==1.5.27
incremental==17.5.0
ipaddr==2.2.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.4.2
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.2
mysql-connector-python==8.0.18
mysqlclient==1.4.2.post1
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.1
oauth2client==4.1.3
opencv-python==4.1.2.30
openpyxl==2.6.3
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.7.0
pep517==0.8.2
pexpect==4.8.0
pg8000==1.13.2
pickleshare==0.7.5
Pillow==7.0.0
pkg-resources==0.0.0
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.5
Protego==0.1.16
protobuf==3.10.0
psutil==5.6.7
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pyarrow==0.17.1
pyasn1==0.4.6
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
PyDispatcher==2.0.5
PyDrive==1.3.1
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
pytesseract==0.3.1
python-dateutil==2.8.0
python-slugify==3.0.3
pytoml==0.1.21
pytz==2019.1
PyVirtualDisplay==0.2.5
PyYAML==5.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
queuelib==1.5.0
requests==2.22.0
retrying==1.3.3
rsa==3.4.2
s3fs==0.4.0
s3transfer==0.3.1
scramp==1.1.0
Scrapy==2.0.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
SQLAlchemy==1.3.15
sqlalchemy-redshift==0.7.9
sqlparse==0.3.0
tenacity==6.0.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
tornado==6.0.4
traitlets==4.3.3
Twisted==19.10.0
typed-ast==1.4.0
unicode==2.7
uritemplate==3.0.1
urllib3==1.25.3
w3lib==1.21.0
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
XlsxWriter==1.2.0
youtube-dl==2020.1.24
zope.interface==4.7.2
